Android DB INTEGER type now returns long value instead of int

+++ src/android/src/main/java/com/futureplatforms/kirin/android/db/AndroidDatabase.java
@@ -210,7 +210,7 @@ public class AndroidDatabase implements DatabaseDelegate {
 							break;
 
 						case FIELD_TYPE_INTEGER: {
-							values.add(String.valueOf(cursor.getInt(i)));
+							values.add(String.valueOf(cursor.getLong(i)));
 						}
 							break;
